Conflicts:
	include/Camera.h
	include/DRGeometrieHeightfield.h
	include/sektor/SubPlanetSektor.h
	src/DRGeometrieHeightfield.cpp
	src/sektor/PlanetSektor.cpp
	src/sektor/SubPatchPlanetSektor.cpp
	src/sektor/SubPlanetSektor.cpp

+++ include/render/RenderNoisePlanetToTexture.h
@@ -14,7 +14,7 @@ public:
     ~RenderNoisePlanetToTexture();
     
     //__inline__ void update(float patchScaling) {mPatchScaling = patchScaling;}
-    DRReturn init(float stepSize, DRVector3 translate, float patchScaling, DRTexturePtr texture, const DRMatrix& rotation = DRMatrix::identity());
+    DRReturn init(int stepSize, DRVector3 translate, float patchScaling, DRTexturePtr texture, const DRMatrix& rotation = DRMatrix::identity());
     
     virtual DRReturn renderStuff();
